refactor: move logic into 'setOnChainTxIdByPayoutId' facade method
